requirements.txt
requests
python-daemon
pyyaml
websocket-client
slackclient
yahoo-finance